Add UploadButton component
