+++ src/Components/Pages/Process/Form/Form.tsx
@@ -1,15 +1,11 @@
 import { Dispatch, useState } from "react";
-import { useHistory } from "react-router-dom";
 import { ConcernShape } from "../../../../interfaces";
-import processTOS, { processTOSPDF } from "../../../../apicalls";
-import Button from "@mui/material/Button";
-import CheckRoundedIcon from "@mui/icons-material/CheckRounded";
-import UploadFileIcon from "@mui/icons-material/UploadFile";
 import { FormTitle } from "./FormTitle";
 import { TOSInput } from "./TOSInput";
 import { ConcersSelector } from "./ConcernsSelector";
 import { TOSLibrary } from "./TOSLibrary";
 import { ProcessButton } from "./ProcessButton";
+import { UploadButton } from "./UploadButton";
 
 interface Props {
   user: number | null;
@@ -24,31 +20,7 @@ interface Props {
 
 export const Form = ({user, tosInput, setLoading, concernAreas, setConcernAreas, setConcerns, setError, setTosInput}: Props) => {
   const [selectedLibrary, setSelectedLibrary] = useState('');
-  const history = useHistory();
 
-  const getConcernCount = () => {
-    const keys = Object.keys(concernAreas);
-    return keys.reduce((acc, key) => {
-      concernAreas[key] && acc++;
-      return acc ;
-    }, 0);
-  };
-
-  const processPDF = async (file: File) => {
-      try {
-        setLoading(true);
-        const concerns = Object.keys(concernAreas).filter(
-          (key) => concernAreas[key]
-        );
-        const TOSinfo = await processTOSPDF(file, concerns, user);
-        setConcerns(TOSinfo.data);
-        setLoading(false);
-        history.push("/results");
-      } catch (error: any) {
-        const errorMessage: string = error.message;
-        errorMessage ? setError(errorMessage) : setError("unexpected error 🙃")
-      };
-  }
   return (
     <form className="form-card">
       <FormTitle />
@@ -58,27 +30,7 @@ export const Form = ({user, tosInput, setLoading, concernAreas, setConcernAreas,
         <TOSLibrary selectedLibrary={selectedLibrary} setSelectedLibrary={setSelectedLibrary} setTosInput={setTosInput}/>
         <div className="buttons-parent">
           <ProcessButton concernAreas={concernAreas} setError={setError} setConcerns={setConcerns} setLoading={setLoading} tosInput={tosInput} user={user}/>
-          <input
-            disabled={!getConcernCount()}
-            accept="application/pdf"
-            style={{ display: "none" }}
-            id="contained-button-file"
-            multiple
-            type="file"
-            onChange={(e) => {
-              if(e.target.files) {
-                processPDF(e.target.files[0])
-              }
-            }}
-          />
-          <label htmlFor="contained-button-file">
-            <Button
-              disabled={!getConcernCount()}
-              component="span"
-              color="primary"
-              variant="outlined"
-              startIcon={<UploadFileIcon />}>Upload</Button>
-          </label>
+          <UploadButton concernAreas={concernAreas} setConcerns={setConcerns} setError={setError} setLoading={setLoading} user={user}/>
         </div>
       </div>
     </form>
